Moved get_token and get_json to a helper file
